Add more to the catalogue, Skip SSL Cert checks (bad idea)
The issue I was facing is that on Android, any Rust crate that tries to
use SSL fails, citing that "The SSL Certificate is expired". I don't
know about Android enough to verify that information, but regardless,
skipping the check fixes it. This should be addressed immediately on the
SSL rust crate's side of things (or maybe the issue lies in the git2
crate, I don't know).